requirements.txt
pytest == 8.0.2
selene>=2.0.0rc9
allure-pytest